Update README.md
I created this project to test and enhance my skills as a React Native developer. I learnt about many skills and techniques while working on this app such as using React Native libraries, using Expo and the Expo Go app, using terminal commands, using APIs and .env to hide key values, amongst many more things. Alongside this I wanted to create a product that I would myself be interested in, and as someone who loves the beach and watersports I thought that creating this application would be a great idea.

I plan to make updates to the app and add more features in the future.